Fix autocmd for *
The default behaviour of the plugin was to map itself to the syntax
event for all files and call its load() function.  However if you have
another plugin that tries to manually configure the plugin (e.g. a
syntax plugin) the firing of this autocmd would happen after the syntax
plugin runs and this would then overwrite the custom configuration with
the default leading to unintended syntax highlighting behaviour.

This commit changes the behaviour for the * case to call a loadIfNone()
function which only loads the default configuration if no configuration
has previously been loaded for the buffer